Moved all current work to fixed-width integers from <stdint.h>

+++ Assembler/hashtable.h
@@ -1,9 +1,11 @@
 #ifndef _HASHTABLE
 #define _HASHTABLE
 
+#include <stdint.h>
+
 struct HashtableEntry {
     char *key;
-    unsigned int item;
+    uint16_t item;
 }
 
 #endif
